pyproject.toml
[project]
authors = [
    { name = "Misael Gonzalez (Rayference)" },
    { name = "Nicolae Marton (Rayference)" },
]
description = "Synthetic scene generator for the ESA DTE-S2GOS service."
name = "s2gos-generator"
requires-python = ">=3.10,<3.12"
version = "0.0.1"
dependencies = [
    "geopandas>=0.13.0",
    "xarray>=2023.1.0", 
    "pyproj>=3.4.0",
    "shapely>=2.0.0",
    "trimesh>=3.20.0",
    "numpy>=1.24.0",
    "pillow>=9.0.0",
    "pyarrow>=10.0.0",
    "eradiate[kernel]",
    "joseki",
    "matplotlib>=3.5.0",
    "rasterio>=1.3.0",
    "rioxarray>=0.13.0",
    "pyyaml>=6.0.0",
]

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["linux-64"]

[tool.pixi.dependencies]
python = ">=3.10,<3.12"

[tool.pixi.pypi-dependencies]
s2gos_generator = { path = ".", editable = true }

[tool.pixi.tasks]

[project.optional-dependencies]
test = ["pytest"]
dev = []
docs = ["mkdocs"]

[tool.pixi.environments]
default = { solve-group = "default" }
dev = { features = ["dev", "test", "docs"], solve-group = "default" }
test = { features = ["test"], solve-group = "default" }

[tool.ruff]
include = ["pyproject.toml", "src/**/*.py", "tests/**/*.py"]

[tool.ruff.lint]
ignore = ["E501", "F821"]
isort = { relative-imports-order = "closest-to-furthest" }
per-file-ignores = { "*.ipynb" = ["E402"] }
select = ["I", "F", "E"]

